Fixed a bug where it would generate too small classes in rare cases
When the class size doesn't need to be rounded it is necessary to add
one. Otherwise the upper bound of the last class will be the smaller
than the largest datum.
Added a check to see if this is the case, and if so add 1 to class_size

If this bug happens again in the future replace that if statement with a
while loop containing the same logic; you should never need to do that.